fix(observability)!: stop duplicating resource identity onto every log record
OTel's LoggerProvider holds one resource for all processors, so dual-mode
logging wraps each batch processor's exporter to stamp its own log.type on
exported records. That wrapper was constructed with the *merged* resource
rather than the delta it exists for.

The consequence was paid on every log line the service emitted: six
attributes -- service.name, service.version, deployment.environment.name and
the telemetry.sdk.{name,language,version} triplet -- were copied onto each
record as record-level duplicates of what the OTLP ResourceLogs.resource
block already ships once per batch, at the cost of wire payload plus a
six-element AddAttributes call per record. The one attribute the wrapper
existed for was never added: records leave logger/otel_bridge.go always
carrying log.type, and the merged resource always declared one, so the
record-wins collision branch dropped it every single time. The six
duplicates were the entire observable effect of the wrapper.

Construct it with the delta instead. createLogResource is deleted, and the
provider's own sdklog.WithResource(res) is now the single place service
identity enters the log pipeline. The wrapper is renamed
processorAttributeExporter (file: processor_attribute_exporter.go) to name
what it is for -- attributes specific to a processor, which the provider's
one shared resource structurally cannot carry -- and its inline filter
buffer shrinks from 16 to 4 now that it sizes a delta rather than a resource.

Record-wins precedence is deliberately unchanged, and is load-bearing twice
over. Dual-mode routing happens at OnEmit, before batching, keyed on the
record's own log.type, so a caller-set value must stay authoritative end to
end. And a record emitted directly through the OTel API by third-party code
carries no log.type at all; it routes to the trace processor, whose wrapper
is what labels it. That injection is why the wrapper survives this change
rather than being deleted outright -- TestUnlabeledRecordReceivesLogTypeStamp
pins it so a future "simplification" cannot quietly remove it.

Behavior change on the OTLP log wire, so it ships with ADR-056 and migration
atom C58.5. Backends that flatten record attributes over resource attributes
see identical values; backends that index the two levels separately stop
matching record-level identity filters until those queries are repointed at
the resource attribute of the same name. Detection is backend-side only --
dashboards, alerts and saved queries -- since no grep over consumer sources
can find it.

This also strengthens ADR-055: with no record-level identity duplicate left,
a flattening backend has nothing to flatten over, and the bridge's
reserved-namespace remap now guards log.type alone. ADR-055 gets a one-line
amendment saying so; its historical narrative is left intact.

Tests: TestExportedRecordCarriesNoIdentityDuplicates is the #914 regression
pin (verified to fail, attributed to its own assertion line, when the
enricher is handed res.Attributes() again).
TestLoggerProviderResourceCarriesServiceIdentity pins the premise the whole
change rests on and passes both before and after it.
TestResourceEnrichmentRestoresIdentityAfterBridgeRemap pinned the removed
backfill and is rewritten as
TestExportedRecordCarriesIdentityOnlyInResourceBlock;
TestBridgeGuardCoversEveryResourceIdentityKey keeps its drift-guard loop with
the record-level assertions inverted and a new resource-block assertion.

Closes #914
